+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Changed:
 - Upgrade to High Availability Dialog Copy
+- NVMe Block Storage banner copy in the Volume Create form
 - Marketplace document link updates and cleanup
 - State/Province field in Billing Contact Info drawer becomes a dropdown if country is US or Canada